+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Support non-string types in query and header parameters
+
 ## [0.0.1] - 2017-10-15
 
 ### Added